[FIX] Replace TSP costs with doubles
Express TSP costs as double instead of int, and add the warehouse as the first stop in the test Main program

+++ src/main/java/domain/service/TSP1.java
@@ -8,8 +8,8 @@ import java.util.Iterator;
 public class TSP1 extends TemplateTSP {
 
 	@Override
-	protected int bound(Integer sommetCourant, Collection<Integer> nonVus) {
-		return 0;
+	protected double bound(Integer sommetCourant, Collection<Integer> nonVus) {
+		return 0.0;
 	}
 
 	@Override
